+++ test/Binder2Spec.hs
@@ -39,38 +39,38 @@ instance MonadNumbering S where
 data Ty
   = Ty'Var (Var S Ty)
   | Ty'Arr Ty Ty
-  | Ty'All (Binder Ty Ty)
+  | Ty'All (Binder Ty S Ty)
 
 data Te
   = Te'Var (Var S Te)
-  | Te'Abs Ty (Binder Te Te)
+  | Te'Abs Ty (Binder Te S Te)
   | Te'App Te Te
-  | Te'Lam (Binder Ty Te)
+  | Te'Lam (Binder Ty S Te)
   | Te'Spe Te Ty
 
 instance MkFree S Ty where
-  mkFree = Ty'Var
+  mkFree = return . Ty'Var
 instance MkFree S Te where
-  mkFree = Te'Var
+  mkFree = return . Te'Var
 
 ty'Var :: Var S Ty -> Box S Ty
 ty'Var = boxVar
 ty'Arr :: Box S Ty -> Box S Ty -> Box S Ty
 ty'Arr a b = Ty'Arr <$> a <*> b
-ty'AllRaw :: Box S (Binder Ty Ty) -> Box S Ty
+ty'AllRaw :: Box S (Binder Ty S Ty) -> Box S Ty
 ty'AllRaw = fmap Ty'All
 ty'All :: Var S Ty -> Box S Ty -> Box S Ty
 ty'All x t = ty'AllRaw $ bind x t
 
 te'Var :: Var S Te -> Box S Te
 te'Var = boxVar
-te'AbsRaw :: Box S Ty -> Box S (Binder Te Te) -> Box S Te
+te'AbsRaw :: Box S Ty -> Box S (Binder Te S Te) -> Box S Te
 te'AbsRaw a f = Te'Abs <$> a <*> f
 te'Abs :: Box S Ty -> Var S Te -> Box S Te -> Box S Te
 te'Abs a x t = te'AbsRaw a $ bind x t
 te'App :: Box S Te -> Box S Te -> Box S Te
 te'App t u = Te'App <$> t <*> u
-te'LamRaw :: Box S (Binder Ty Te) -> Box S Te
+te'LamRaw :: Box S (Binder Ty S Te) -> Box S Te
 te'LamRaw = fmap Te'Lam
 te'Lam :: Var S Ty -> Box S Te -> Box S Te
 te'Lam x t = te'LamRaw $ bind x t
@@ -88,36 +88,41 @@ boxTe (Te'App t u) = te'App <$> boxTe t <*> boxTe u
 boxTe (Te'Lam f) = te'LamRaw <$> boxBinder boxTe f
 boxTe (Te'Spe t a) = te'Spe <$> boxTe t <*> boxTy a
 
-hnf :: Te -> Te
-hnf (Te'App t u) = let v = hnf u in case hnf t of
-  Te'Abs _ b -> hnf $ subst b v
-  h -> Te'App h v
-hnf (Te'Spe t a) = case hnf t of
-  Te'Lam b -> hnf $ subst b a
-  h -> Te'Spe h a
-hnf t = t
+hnf :: Te -> S Te
+hnf (Te'App t u) = do
+  hu <- hnf u
+  ht <- hnf t
+  case ht of
+    Te'Abs _ b -> hnf =<< subst b hu
+    _ -> return $ Te'App ht hu
+hnf (Te'Spe t a) = do
+  ht <- hnf t
+  case ht of
+    Te'Lam b -> hnf =<< subst b a
+    _ -> return $ Te'Spe ht a
+hnf t = return t
 
 nf :: Te -> S Te
 nf (Te'Abs a f) = do
   (x, t) <- unbind f
   nt <- nf t
   bt <- boxTe nt
-  return $ Te'Abs a $ unbox $ bind x bt
+  fmap (Te'Abs a) $ unbox $ bind x bt
 nf (Te'App t u) = do
   nt <- nf t
   nu <- nf u
   case nt of
-    Te'Abs _ f -> nf $ subst f u
+    Te'Abs _ f -> nf =<< subst f u
     _ -> return $ Te'App nt nu
 nf (Te'Lam f) = do
   (x, t) <- unbind f
   nt <- nf t
   bt <- boxTe nt
-  return $ Te'Lam $ unbox $ bind x bt
+  fmap Te'Lam $ unbox $ bind x bt
 nf (Te'Spe t a) = do
   nt <- nf t
   case nt of
-    Te'Lam f -> nf $ subst f a
+    Te'Lam f -> nf =<< subst f a
     _ -> return $ Te'Spe nt a
 nf t = return t
 
@@ -155,12 +160,12 @@ infer ctxt (Te'Lam f) = do
   case mtyt of
     Just tyt -> do
       bt <- boxTy tyt
-      return $ Just $ Ty'All $ unbox $ bind x bt
+      fmap (Just . Ty'All) $ unbox $ bind x bt
     Nothing -> return Nothing
 infer ctxt (Te'Spe t a) = do
   mtyt <- infer ctxt t
   case mtyt of
-    Just (Ty'All f) -> return $ Just $ subst f a
+    Just (Ty'All f) -> Just <$> subst f a
     _ -> return Nothing
 
 check :: Ctxt -> Te -> Ty -> S Bool
@@ -206,19 +211,19 @@ term1 :: S Te
 type1 = do
   x <- newVar "X"
   y <- newVar "Y"
-  return $ unbox $ ty'Arr (ty'Var x) (ty'Var y)
+  unbox $ ty'Arr (ty'Var x) (ty'Var y)
 type2 = do
   x <- newVar "X"
   y <- newVar "Y"
   let arr = ty'Arr (ty'Var x) (ty'Var y)
-  return $ unbox $ ty'All x $ ty'All y $ ty'Arr arr arr
+  unbox $ ty'All x $ ty'All y $ ty'Arr arr arr
 term1 = do
   x <- newVar "X"
   y <- newVar "Y"
   f <- newVar "f"
   a <- newVar "a"
   let arr = ty'Arr (ty'Var x) (ty'Var y)
-  return $ unbox $ te'Lam x $ te'Lam y $ te'Abs arr f $ te'Abs (ty'Var x) a $
+  unbox $ te'Lam x $ te'Lam y $ te'Abs arr f $ te'Abs (ty'Var x) a $
     te'App (te'Var f) (te'Var a)
 
 spec :: Spec