+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## version 0
 
+### 0.2 -- still not released
+
+* Now MkFree requires monadic instance.
+* Binder requires Monad typeclass.
+
 ### 0.1 -- 2023-10-04
 
 Initial release.